flask/requirements.txt
opentelemetry-instrumentation-asyncio==0.52b0
opentelemetry-instrumentation-dbapi==0.52b0
opentelemetry-instrumentation-logging==0.52b0
opentelemetry-instrumentation-sqlite3==0.52b0
opentelemetry-instrumentation-threading==0.52b0
opentelemetry-instrumentation-urllib==0.52b0
opentelemetry-instrumentation-wsgi==0.52b0
opentelemetry-instrumentation-aiohttp-client==0.52b0
opentelemetry-instrumentation-aiohttp-server==0.52b0
opentelemetry-instrumentation-click==0.52b0
opentelemetry-instrumentation-flask==0.52b0
opentelemetry-instrumentation-grpc==0.52b0
opentelemetry-instrumentation-httpx==0.52b0
opentelemetry-instrumentation-jinja2==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-instrumentation-sqlalchemy==0.52b0
opentelemetry-instrumentation-system-metrics==0.52b0
opentelemetry-instrumentation-tornado==0.52b0
opentelemetry-instrumentation-tortoiseorm==0.52b0
opentelemetry-instrumentation-urllib3==0.52b0